Crate discord_base

Modules

cmd_error
The module for error handling of the commands

cmd_help
The module for the help command

cmd_info
The module for the info command

cmd_prefix
The module for the prefix command

globals
The module for the statics and structs to save the statics to

Structs

General
The group for the commands provided by default

Handler
The event handler struct to implement EventHandler for

Master
The hidden group for all the commands to be added to

Statics

GENERAL_GROUP
The group for the commands provided by default

GENERAL_GROUP_OPTIONS
The group for the commands provided by default

MASTER_GROUP
The hidden group for all the commands to be added to

MASTER_GROUP_OPTIONS
The hidden group for all the commands to be added to

Functions

log
DMs the owner of the bot, as in the application info, with the message

print_and_write
Prints the msg and the timestamp and appends it (or creates if it doesn't exist) to the log file in the config

send_embed

Sets the colour of the embed to 11534368 (The baseline error colour according to Material Design guidelines) if is_error is true, if not, sets it to the colour in the config

Sends the embed to the channel_id of reply

set_dir

Sets the working directory to the directory of the binary, so that the config file and all are saved to the same directory as the file, as expected

Also prints the working directory, just for info