Function discord_base::send_embed

pub async fn send_embed(
    ctx: &Context, 
    reply: &Message, 
    is_error: bool, 
    embed: CreateEmbed
)

Sets the colour of the embed to 11534368 (The baseline error colour according to Material Design guidelines) if is_error is true, if not, sets it to the colour in the config

Sends the embed to the channel_id of reply

Error

Uses log() to inform why setting the colour failed and falls back to the Default colour (most likely white)

Says why it couldn't send the embed in the channel in plain text (without embeds)

DMs the author of reply if that also fails, colouring the embed with the error colour and telling them to report to the admins

Uses log() to inform why it failed if even that fails