Function discord_base::set_dir

pub fn set_dir()

Sets the working directory to the directory of the binary, so that the config file and all are saved to the same directory as the file, as expected

Also prints the working directory, just for info

Error

Prints why it couldn't change the directory

Doesn't panic since the program can still run in the other directory which will be printed